+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG for Sulu Product Bundle
 
 * dev-develop
 
+    * FEATURE     Implemented new variants api to create, update and delete product variants.
+    * FEATURE     Added product type key to id map as parameter to container.
     * FEATURE     Added validation schema for get products api.
 
 * 0.15.0 (2016-09-26)